move test generation into seperate file
